랩핑 추가 및 app.py에서 모듈화 호출 방식으로 변경
